+++ app/src/main/java/eu/faircode/email/ApplicationEx.java
@@ -46,6 +46,10 @@ import androidx.core.os.LocaleListCompat;
 import androidx.emoji2.text.DefaultEmojiCompatConfig;
 import androidx.emoji2.text.EmojiCompat;
 import androidx.emoji2.text.FontRequestEmojiCompatConfig;
+import androidx.lifecycle.Lifecycle;
+import androidx.lifecycle.LifecycleObserver;
+import androidx.lifecycle.OnLifecycleEvent;
+import androidx.lifecycle.ProcessLifecycleOwner;
 import androidx.preference.PreferenceManager;
 import androidx.work.WorkManager;
 
@@ -125,6 +129,22 @@ public class ApplicationEx extends Application
             UriHelper.test(this);
 
         CoalMine.install(this);
+
+        ProcessLifecycleOwner.get().getLifecycle().addObserver(new LifecycleObserver() {
+            @OnLifecycleEvent(Lifecycle.Event.ON_START)
+            public void onStart() {
+                log(true);
+            }
+
+            @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
+            public void onStop() {
+                log(false);
+            }
+
+            private void log(boolean foreground) {
+                Log.breadcrumb("app", "foreground", Boolean.toString(foreground));
+            }
+        });
 
         registerActivityLifecycleCallbacks(lifecycleCallbacks);
 
@@ -741,6 +761,9 @@ public class ApplicationEx extends Application
         } else if (version < 2089) {
             if (!prefs.contains("auto_hide_answer"))
                 editor.putBoolean("auto_hide_answer", !Helper.isAccessibilityEnabled(context));
+        } else if (version < 2108) {
+            if (!prefs.getBoolean("updown", false))
+                editor.putBoolean("updown", false);
         }
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && !BuildConfig.DEBUG)