+++ app/src/main/assets/CHANGELOG.md
@@ -6,6 +6,33 @@ For support you can use [the contact form](https://contact.faircode.eu/?product=
 
 ### [Vallibonavenatrix](https://en.wikipedia.org/wiki/Vallibonavenatrix)
 
+### 1.2110 - 2023-10-14
+
+* Added PDF thumbnail preview (mini gallery)
+* Small improvements and minor bug fixes
+* Updated [translations](https://crowdin.com/project/open-source-email)
+
+### 1.2109 - 2023-10-14
+
+* Small improvements and minor bug fixes
+
+### 1.2108 - 2023-10-12
+
+* Added [Liberation Sans Narrow](https://github.com/liberationfonts/liberation-sans-narrow) font (advanced display options)
+* Added *$weekday$* placeholder for templates
+* Added option to disabling printing of block quotes lines
+* Added accessibility actions to mark a message read/unread and to trash a conversation
+* Small improvements and minor bug fixes
+* Updated [AndroidX](https://developer.android.com/jetpack/androidx/versions/all-channel)
+* Updated [translations](https://crowdin.com/project/open-source-email)
+
+### 1.2107 - 2023-10-04
+
+* Added rule action to execute webhooks
+* Small improvements and minor bug fixes
+* Updated [Public Suffix List](https://github.com/publicsuffix/list)
+* Updated [translations](https://crowdin.com/project/open-source-email)
+
 ### 1.2106 - 2023-09-27
 
 * Added option to use up/down arrows for next/previous conversation